Prid refactor (#16)
# Goal
The goal of this PR is to make sure that Prids are compatible with the
test vector

related to  #9 

# Discussion
- It seems like that for **prid a->b** we were encrypting `id_a` as the
message instead of `id_b` and using the other one as nonce. It does seem
like in the early versions of the algorithm (at least mentioned in
slack) we were doing this but we switched to the other way around. After
switching to the latest version of the algorithm we are compatible with
the [test
vector](https://github.com/LibertyDSNP/prid-ts/blob/main/index.test.ts)
- Disabling clippy for now to speed up development process and we will
enable and address comments when repo is in a more stable state
- moving other tests from their implementation files into their own
file. We will need to have a decision about where is the best location
for tests.
- used `u64` for `lastUpdated` since zigzag encoding takes care of
efficiency and we don't have any other u32 in schema

# Checklist
- [X] Tests added

+++ core/src/types.rs
@@ -9,7 +9,7 @@ pub struct PrivateGraphChunk {
 	pub prids: Vec<DsnpPrid>,
 
 	/// Days since the Unix Epoch when PRIds for this chunk were last refreshed
-	pub last_updated: u32,
+	pub last_updated: u64,
 
 	/// connections
 	pub inner_graph: DsnpInnerGraph,
